Done Users and Items saving by repository

+++ src/main/java/ru/practicum/shareit/user/UserController.java
@@ -28,7 +28,7 @@ public class UserController {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<UserDto> getUser(@PathVariable(required = false) int id) {
+    public ResponseEntity<UserDto> getUser(@PathVariable(required = false) long id) {
         return new ResponseEntity<>(userService.getUser(id), HttpStatus.OK); //вернет пользователя или исключение
     }
 
@@ -38,13 +38,13 @@ public class UserController {
     }
 
     @PatchMapping("/{id}")
-    public ResponseEntity<UserDto> updateUser(@PathVariable int id,
+    public ResponseEntity<UserDto> updateUser(@PathVariable long id,
                                               @RequestBody UserUpdateDto userUpdateDto) {
         return new ResponseEntity<>(userService.update(userUpdateDto, id), HttpStatus.OK);
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<String> deleteUser(@PathVariable int id) {
+    public ResponseEntity<String> deleteUser(@PathVariable long id) {
         return new ResponseEntity<>(userService.delete(id), HttpStatus.OK);
     }
 }
